+++ projects/qgrid-plugins/import/import.plugin.d.ts
@@ -2,14 +2,11 @@ import { Model } from '@qgrid/core/model/model';
 import { Command } from '@qgrid/core/command/command';
 
 export declare interface ImportOptions {
-	head?: 'alpha' | 'numeric' | 'default';
+	head: 'alpha' | 'numeric' | 'default';
 }
 export declare class ImportPlugin {
-	constructor(model: Model, context: { element: HTMLElement, options?: ImportOptions });
-
-	model: { [ key: string ]: any };
-	options?: ImportOptions;
-	upload: Command;
+	constructor(model: Model, element: HTMLElement, options?: ImportOptions);
 
 	load(e: any): void;
+	upload(): void;
 }